+++ berkeley-function-call-leaderboard/bfcl/eval_checker/model_metadata.py
@@ -497,6 +497,30 @@ MODEL_METADATA_MAPPING = {
         "Microsoft",
         "MIT",
     ],
+    "Qwen/Qwen2-1.5B-Instruct": [
+        "Qwen2-1.5B-Instruct (Prompt)",
+        "https://huggingface.co/Qwen/Qwen2-1.5B-Instruct",
+        "Qwen",
+        "apache-2.0",
+    ],
+    "Qwen/Qwen2-7B-Instruct": [
+        "Qwen2-7B-Instruct (Prompt)",
+        "https://huggingface.co/Qwen/Qwen2-7B-Instruct",
+        "Qwen",
+        "apache-2.0",
+    ],
+    "Qwen/Qwen2.5-1.5B-Instruct": [
+        "Qwen2.5-1.5B-Instruct (Prompt)",
+        "https://huggingface.co/Qwen/Qwen2.5-1.5B-Instruct",
+        "Qwen",
+        "apache-2.0",
+    ],
+    "Qwen/Qwen2.5-7B-Instruct": [
+        "Qwen2.5-7B-Instruct (Prompt)",
+        "https://huggingface.co/Qwen/Qwen2.5-7B-Instruct",
+        "Qwen",
+        "apache-2.0",
+    ],
 }
 
 INPUT_PRICE_PER_MILLION_TOKEN = {